requirements.txt
# Django
Django==4.1

# Decouple
python-decouple==3.6

# Heroku
dj-database-url==1.0.0
django-heroku==0.3.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
gunicorn==20.1.0

# Images
Pillow==9.2.0

# Cloudinary
dj3-cloudinary-storage==0.0.6

# Bling
pybling==0.0.9

# Sentry
sentry-sdk==1.9.6

# Crispy Forms
django-crispy-forms==1.14.0

# Bulk Changes in Admin
django-mass-edit==3.4.1

# Excel for reports
xlwt==1.3.0

# PDF
xhtml2pdf==0.2.8

# Loggin
django-simple-history==3.1.1

# Analytics
django-analytical==3.1.0
